requirements.txt
flask
flask-SQLAlchemy
python-dotenv   
Werkzeug